requirements.txt
aiohttp>=3.9
httpx>=0.27
PyYAML>=6.0
croniter>=2.0
